\name{gse1992}
\alias{gse1992}
\docType{data}
\title{
GSE1992
}
\description{
Gene expression data of GSE1992. Probes/genes mapping was performed via the CleanEx database, release of September 2007.
}
\usage{data(gse1992)}
\format{
  A data frame with 124 observations on the 15528 numeric variables.
}
\source{
\url{http://www.ncbi.nlm.nih.gov/projects/geo/query/acc.cgi?acc=GSE1992}
}
\references{
 Yasrebi H,  Sperisen P,  Praz V,  Bucher P, 2009 Can Survival Prediction Be Improved By Merging Gene Expression Data Sets?. PLoS ONE 4(10): e7431. doi:10.1371/journal.pone.0007431  
}
\seealso{
\code{\link{gse1992pheno}} 
}


\examples{
data(gse1992)
}
\keyword{datasets}
