\name{gse3143}
\alias{gse3143}
\docType{data}
\title{
GSE3143
}
\description{
Gene expression data of GSE3143. Probsets/genes mapping was performed via the CleanEx database, release of September 2007.
}
\usage{data(gse3143)}
\format{
  A data frame with 158 observations on the 8659 numeric variables.
  }

\source{
\url{http://www.ncbi.nlm.nih.gov/projects/geo/query/acc.cgi?acc=GSE3143}

}
\references{
 Yasrebi H,  Sperisen P,  Praz V,  Bucher P, 2009 Can Survival Prediction Be Improved By Merging Gene Expression Data Sets?. PLoS ONE 4(10): e7431. doi:10.1371/journal.pone.0007431.  
}
\seealso{
\code{\link{gse3143pheno}} 
}

\examples{
data(gse3143)
}
\keyword{datasets}
