\name{gse4335}
\alias{gse4335}
\docType{data}
\title{
GSE4335
}
\description{
Gene expression data of GSE4335. Probes/genes mapping was performed via the CleanEx database, release of September 2007.
}
\usage{data(gse4335)}
\format{
  A data frame with 115 observations on the 12793 numeric variables.
}

\source{
\url{http://www.ncbi.nlm.nih.gov/projects/geo/query/acc.cgi?acc=GSE4335}
}
\references{
 Yasrebi H,  Sperisen P,  Praz V,  Bucher P, 2009 Can Survival Prediction Be Improved By Merging Gene Expression Data Sets?. PLoS ONE 4(10): e7431. doi:10.1371/journal.pone.0007431  
}
\seealso{
\code{\link{gse4335pheno}}
}

\examples{
data(gse4335)
}
\keyword{datasets}
