\name{survJamda.data-package}
\alias{survJamda.data-package}
\alias{survJamda.data}
\docType{package}
\title{
Three breast cancer data sets for survJamda-R package
}
\description{
Three microarray gene expression breast cancer data sets that can be used with the survJamda
package
}
\details{
\tabular{ll}{
Package: \tab survJamda.data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Haleh Yasrebi

Maintainer: Haleh Yasrebi hyasrebi@yahoo.com
}
\references{
Yasrebi H, Sperisen P, Praz V, Bucher P, 2009 Can Survival Prediction Be Imp roved By Merging
Gene Expression Data Sets?. PLoS ONE 4(10): e7431. doi:10.1371/journal.pone.0007431
}
\keyword{ package }
\seealso{
 \code{\link[survJamda:survJamda-package]{survJamda}} ~~
}
